[ refactor ] Header ~ ProgressBar 통합
