+++ src/@components/@common/SubHeader/style.ts
@@ -14,4 +14,16 @@ export const St = {
   GoBackBtnContainer: styled.div``,
 
   CloseBtnContainer: styled.div``,
+
+  ProgressBarContainer: styled.section`
+    width: 100%;
+    height: 0.4rem;
+  `,
+
+  ProgressRate: styled.div<{ rate: number }>`
+    height: 100%;
+    width: ${({ rate }) => rate}%;
+
+    background-color: ${({ theme }) => theme.colors.green};
+  `,
 };
